+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Handle early return from `process_graphql_query` in `py_gql.graphql`.
 - Make sure `process_graphql_query` calls `tracer.on_end()` on early returns.
 - Fix link in the description of the `UUID` scalar type.
+- Fix bug in `Executor.complete_value` which lead to incorrectly ignoring fragment when applied on types implementing an interface.
 
 ## 0.2.0